make testcase logfile timstamps equidistant
The created and handcrafted logfiles can't be replayed on a real
CAN interface as the handcrafted timestamps only take care of the
order and not whether the CAN frame could be sent on the wire.
Therefore these logfiles have been reworked to enable the (time
consuming) sending of the content on a real CAN interface.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>